+++ CHANGELOG.md
@@ -8,6 +8,15 @@
   be used with TOP_RESULTS query option and sorting term specified but not
   limited to.
 
+* Breaking change: Gson's '@SerializedName(value)' is now properly supported,
+  the field name exposed in the DocumentDescription and available in xenon
+  queries now respects Gson's @SerializedName annotation.
+
+  Previously, the exposed JSON document respected this annotation but the
+  DocumentDescription and query service did not respect the annotation,
+  exposing the underlying field with its java name, which was incoherent.
+
+  See: https://www.pivotaltracker.com/story/show/148644281  
 
 ## 1.5.1